# Fix segfault of #877