This fixes the segfault on Julia v1.0 of https://github.com/JuliaOpt/MathOptInterface.jl/pull/877

This also give an impressive performance improvements for the benchmark added in https://github.com/JuliaOpt/MathOptInterface.jl/pull/946. This performance improvement might be explained by https://github.com/JuliaLang/julia/issues/32761
On master:
```
BenchmarkTools.Trial: 
  memory estimate:  269.80 KiB
  allocs estimate:  4658
  --------------
  minimum time:     11.074 ms (0.00% GC)
  median time:      11.251 ms (0.00% GC)
  mean time:        11.761 ms (1.45% GC)
  maximum time:     61.752 ms (79.61% GC)
  --------------
  samples:          425
  evals/sample:     1
BenchmarkTools.Trial: 
  memory estimate:  715.16 KiB
  allocs estimate:  12004
  --------------
  minimum time:     41.166 ms (0.00% GC)
  median time:      45.770 ms (0.00% GC)
  mean time:        50.720 ms (2.82% GC)
  maximum time:     203.532 ms (60.85% GC)
  --------------
  samples:          99
  evals/sample:     1
```
After this PR:
```
BenchmarkTools.Trial: 
  memory estimate:  68.59 KiB
  allocs estimate:  1861
  --------------
  minimum time:     3.819 ms (0.00% GC)
  median time:      5.165 ms (0.00% GC)
  mean time:        5.701 ms (2.53% GC)
  maximum time:     110.109 ms (91.74% GC)
  --------------
  samples:          875
  evals/sample:     1
BenchmarkTools.Trial: 
  memory estimate:  163.96 KiB
  allocs estimate:  4440
  --------------
  minimum time:     23.947 ms (0.00% GC)
  median time:      32.318 ms (0.00% GC)
  mean time:        32.113 ms (0.80% GC)
  maximum time:     83.834 ms (47.83% GC)
  --------------
  samples:          156
  evals/sample:     1
```