@frapac 2019-11-07
Just being curious: the segfault was induced by the splashing `args..`?