@blegat 2019-11-07
> Just being curious: the segfault was induced by the splashing `args..`?

The only thing I know is that calling the function caused a segfault and calling the function it was calling was working. So I tried simplifying the work for the compiler by removing the splatting. In fact, I recently discovered that splatting shouldn't be used for performance-critical function https://github.com/JuliaLang/julia/issues/32761, this should be in the performance tips.