@frapac 2019-11-07
Interesting. I remember having a lot of problems with splatting in Julia. Indeed, that should be specified in the performance tips...